Game 1 basics working
Plotted numbers are alway positive. This must be fixed.

+++ KoorweekendApp2017/Models/ChoirWeekendBasePosition.cs
@@ -13,9 +13,9 @@ namespace KoorweekendApp2017.Models
     public class ChoirWeekendBasePosition
     {
 		
-		public float Longitude { get; set; }
+		public double Longitude { get; set; }
 
-		public float Lattitude { get; set; }
+		public double Lattitude { get; set; }
 
     }
 }
